Agent Manager Refactor
Created an agent manager server (and later, agent manager script) that is a socket client. This server gets paired with a web client via the socket server. There is messaging set up to forward messages between the web client and agent manager's server.js.

The video parser has been migrated into the agent manager server (and later, will be moveed into the agent manager script).

Currently, a message from the user is forwarded to the server and a "request received" type message it sent back to the web user from the connected server.

This server and associated scripts will soon handle the LLM functionality and majority of the workflows integrated.